+++ dex-test-common/build.sbt
@@ -1,0 +1,3 @@
+description := "Shared classes for unit tests of DEX and waves-integration"
+
+libraryDependencies ++= Dependencies.testCommon.value
