+++ waves-integration-it/src/test/resources/logback-test.xml
@@ -23,6 +23,7 @@
 
     <logger name="org.asynchttpclient" level="INFO"/>
     <logger name="io.netty" level="INFO"/>
+    <logger name="io.grpc.netty" level="WARN"/>
     <logger name="io.swagger" level="OFF"/>
 
     <root level="TRACE">
