--- waves-integration/src/main/protobuf/waves_balances_api.proto
@@ -1,19 +1,0 @@
-syntax = "proto3";
-
-package waves.dex.grpc.integration;
-option java_package = "com.wavesplatform.dex.grpc.integration.services";
-
-import "google/protobuf/empty.proto";
-
-service WavesBalancesApi {
-    rpc GetBalanceChanges (google.protobuf.Empty) returns (stream BalanceChangesResponse);
-}
-
-message BalanceChangesResponse {
-    message Record {
-        bytes address = 1;
-        bytes asset = 2;
-        int64 balance = 3;
-    }
-    repeated Record batch = 1;
-}
