+++ waves-integration/src/main/scala/com/wavesplatform/dex/grpc/integration/DEXClient.scala
@@ -2,8 +2,7 @@ package com.wavesplatform.dex.grpc.integration
 
 import java.util.concurrent.TimeUnit
 
-import com.wavesplatform.dex.grpc.integration.clients.async.WavesBlockchainCachingClient
-import com.wavesplatform.dex.grpc.integration.clients.sync.WavesBlockchainGrpcSyncClient
+import com.wavesplatform.dex.grpc.integration.clients.{WavesBlockchainCachingClient, WavesBlockchainGrpcAsyncClient}
 import com.wavesplatform.utils.ScorexLogging
 import io.grpc._
 import io.grpc.internal.DnsNameResolverProvider
@@ -12,7 +11,8 @@ import monix.execution.Scheduler
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration.FiniteDuration
 
-class DEXClient(target: String, defaultCachesExpiration: FiniteDuration, val monixScheduler: Scheduler, val grpcExecutionContext: ExecutionContext)
+class DEXClient(target: String, defaultCachesExpiration: FiniteDuration)(implicit val monixScheduler: Scheduler,
+                                                                         val grpcExecutionContext: ExecutionContext)
     extends ScorexLogging {
 
   log.info(s"NODE gRPC server: $target")
@@ -30,6 +30,8 @@ class DEXClient(target: String, defaultCachesExpiration: FiniteDuration, val mon
       .usePlaintext()
       .build
 
-  lazy val wavesBlockchainSyncClient  = new WavesBlockchainGrpcSyncClient(channel)
-  lazy val wavesBlockchainAsyncClient = new WavesBlockchainCachingClient(channel, defaultCachesExpiration, monixScheduler)(grpcExecutionContext)
+  lazy val wavesBlockchainAsyncClient = new WavesBlockchainCachingClient(
+    new WavesBlockchainGrpcAsyncClient(channel),
+    defaultCachesExpiration
+  )
 }
