+++ waves-integration/src/main/scala/com/wavesplatform/dex/grpc/integration/clients/RunScriptResult.scala
@@ -1,0 +1,10 @@
+package com.wavesplatform.dex.grpc.integration.clients
+
+sealed trait RunScriptResult
+object RunScriptResult {
+  case class ScriptError(message: String)             extends RunScriptResult
+  case object Denied                                  extends RunScriptResult
+  case object Allowed                                 extends RunScriptResult
+  case class UnexpectedResult(rawResult: String)      extends RunScriptResult
+  case class Exception(name: String, message: String) extends RunScriptResult
+}
