+++ waves-integration/src/main/scala/com/wavesplatform/dex/grpc/integration/protobuf/package.scala
@@ -1,6 +1,6 @@
 package com.wavesplatform.dex.grpc.integration
 
-import java.util.concurrent.atomic.AtomicReference
+import java.util.concurrent.LinkedBlockingQueue
 
 import com.wavesplatform.api.http.ApiError
 import com.wavesplatform.lang.ValidationError
@@ -9,26 +9,25 @@ import monix.eval.Task
 import monix.execution.{Cancelable, Scheduler}
 import monix.reactive.Observable
 
-package object protobuf {
+import scala.util.control.NonFatal
 
+package object protobuf {
   implicit class StreamObserverMonixOps[T](streamObserver: StreamObserver[T])(implicit sc: Scheduler) {
     // TODO: More convenient back-pressure implementation
     def toSubscriber: monix.reactive.observers.Subscriber[T] = {
       import org.reactivestreams.{Subscriber, Subscription}
 
       val rxs = new Subscriber[T] with Cancelable {
-        private[this] val element = new AtomicReference[Option[T]](None)
+        private[this] val queue = new LinkedBlockingQueue[T](32)
 
         @volatile
         private[this] var subscription: Subscription = _
 
         private[this] val observerReadyFunc: () => Boolean = streamObserver match {
           case callStreamObserver: CallStreamObserver[_] =>
-            () =>
-              callStreamObserver.isReady
+            () => callStreamObserver.isReady
           case _ =>
-            () =>
-              true
+            () => true
         }
 
         def isReady: Boolean = observerReadyFunc()
@@ -36,14 +35,11 @@ package object protobuf {
         override def onSubscribe(subscription: Subscription): Unit = {
           this.subscription = subscription
 
-          def pushElement(): Unit = element.get() match {
-            case v @ Some(value) if this.isReady =>
-              if (element.compareAndSet(v, None)) {
-                streamObserver.onNext(value)
-                subscription.request(1)
-              } else {
-                pushElement()
-              }
+          def pushElement(): Unit = Option(queue.peek()) match {
+            case Some(_) if this.isReady =>
+              val qv = queue.poll()
+              streamObserver.onNext(qv)
+              subscription.request(1)
 
             case None if this.isReady =>
               subscription.request(1)
@@ -57,12 +53,10 @@ package object protobuf {
               scso.disableAutoInboundFlowControl()
               scso.setOnCancelHandler(() => subscription.cancel())
               scso.setOnReadyHandler(() => pushElement())
-            // subscription.request(1)
 
             case cso: CallStreamObserver[T] =>
               cso.disableAutoInboundFlowControl()
               cso.setOnReadyHandler(() => pushElement())
-            // subscription.request(1)
 
             case _ =>
               subscription.request(Long.MaxValue)
@@ -70,17 +64,11 @@ package object protobuf {
         }
 
         override def onNext(t: T): Unit = {
+          queue.add(t)
           if (isReady) {
-            val value = element.get()
-            if (value.nonEmpty) {
-              if (element.compareAndSet(value, Some(t))) streamObserver.onNext(value.get)
-              else onNext(t)
-            } else {
-              streamObserver.onNext(t)
-            }
+            val value = Option(queue.poll())
+            value.foreach(streamObserver.onNext)
             if (isReady) subscription.request(1)
-          } else if (!element.compareAndSet(None, Some(t))) {
-            throw new IllegalArgumentException("Buffer overflow")
           }
         }
 
@@ -108,6 +96,10 @@ package object protobuf {
     def failWith(error: ApiError): Unit = {
       streamObserver.onError(GRPCErrors.toStatusException(error))
     }
+
+    def interceptErrors(f: => Unit): Unit =
+      try f
+      catch { case NonFatal(e) => streamObserver.onError(GRPCErrors.toStatusException(e)) }
   }
 
   implicit class EitherVEExt[T](e: Either[ValidationError, T]) {
